big organisation modification + split files + clear merge sort with specified compare function

+++ srcs/new/new_infos.c
@@ -6,7 +6,7 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/26 20:38:16 by rpinoit           #+#    #+#             */
-/*   Updated: 2018/09/12 21:18:48 by rpinoit          ###   ########.fr       */
+/*   Updated: 2018/09/15 11:42:26 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,12 +19,12 @@ t_infos  *new_infos(struct stat *sd)
     if ((new = (t_infos *)malloc(sizeof(t_infos))) == NULL)
         error_malloc();
     new->mode = infos_mode(sd->st_mode);
-    new->link = (int)sd->st_nlink;
+    new->link = (unsigned long)sd->st_nlink;
     new->uid = infos_uid(sd->st_uid);
     new->gid = infos_gid(sd->st_gid);
-    new->size = (int)sd->st_size;
-    new->blocks = (int)sd->st_blocks;
-    new->mtime = (int)sd->st_mtime;
+    new->size = (unsigned long)sd->st_size;
+    new->blocks = (unsigned long)sd->st_blocks;
+    new->mtime = (unsigned long)sd->st_mtime;
     new->time = infos_time(sd->st_mtime);
     return (new);
 }
